requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9
pytest==8.4.2
httpx==0.28.1
PyJWT==2.10.1
alembic==1.16.5
